requirements.txt

openai
chromadb
python-dotenv
pydantic
pydantic-settings
python-multipart
pypdf
fastapi
uvicorn
streamlit
requests
pymongo
passlib
sentence-transformers
bcrypt
reportlab
